game,interpreter,main: bye bye io library 👋
The game engine doesn't have support for the `io` library since it tries its
best to hide the underlying filesystem so things could be nicely abstracted.
It instead only provides a function `lua.ReadFile()`, which reads all file
contents at once and returns them as a string. Oh, and fortunately it
also kept the string library so I wouldn't have to go all low-level in
the code.
So there it is, I changed the elegant `io` file parsing calls to slightly
less readable regexes.