non-blocking maze drawing, memory savement
The JudgeWrapper actor now gets a hint about when to refresh
currently displayed data, but queues this hint for themselves so they're
not slowing down the draw loop with data preparation.

This actor also no longer creates a new table each time they're
preparing a new screen which simplifies the code a bit.